style(ifu): move prefixOr function to utils package
